+++ exhibitera/apps/InfoStation/infostation.js
@@ -36,6 +36,8 @@ function loadDefinition (definition) {
   // Layout
   root.style.setProperty('--button-size', definition?.style?.layout?.button_size || 30)
   root.style.setProperty('--header-height', definition?.style?.layout?.header_height || 10)
+  root.style.setProperty('--sidebar-width', definition?.style?.layout?.sidebar_width || 30)
+  root.style.setProperty('--toolbar-height', definition?.style?.layout?.toolbar_height || 5)
 
   if (parseInt(definition?.style?.layout?.header_height || 10) === 0) {
     document.getElementById('mastheadDiv').style.display = 'none'